docs: oppdater fargedokumentasjonen
affects: @fremtind/portal

ISSUES CLOSED: #2417
