--- portal/src/components/Documentation/Colors/ColorDescription.scss
@@ -1,103 +1,0 @@
-@use "~@fremtind/jkl-core/jkl";
-
-@import "~@fremtind/jkl-core/mixins/_all.scss";
-@import "~@fremtind/jkl-core/_functions.scss";
-
-.jkl-portal-color-description {
-    width: 100%;
-    max-width: rem(768px);
-    margin-top: jkl.$spacing-3xl;
-
-    @include small-device {
-        &__title,
-        &__swatch {
-            margin-bottom: jkl.$spacing-xl;
-        }
-    }
-
-    @include from-medium-device {
-        display: grid;
-        grid-template-columns: rem(340px) 1fr;
-
-        &__title {
-            grid-column: 2 / 3;
-        }
-    }
-}
-
-.jkl-portal-two-tone-description {
-    width: 100%;
-    max-width: rem(768px);
-
-    &__title {
-        margin-top: jkl.$spacing-2xl;
-        margin-bottom: jkl.$spacing-s;
-    }
-
-    &__info {
-        margin-bottom: jkl.$spacing-xl;
-    }
-
-    &__light-color,
-    &__dark-color {
-        position: relative;
-        padding: jkl.$spacing-2xl 0;
-    }
-
-    &__light-color:before,
-    &__dark-color:before {
-        @include jkl.text-style("small");
-        position: absolute;
-        top: jkl.$spacing-s;
-    }
-
-    &__light-color {
-        color: jkl.$color-granitt;
-        background-color: jkl.$color-snohvit;
-
-        &:before {
-            content: "Lys bakgrunn";
-            color: jkl.$color-stein;
-        }
-    }
-
-    &__dark-color {
-        color: jkl.$color-snohvit;
-        background-color: jkl.$color-granitt;
-
-        &:before {
-            content: "Mørk bakgrunn";
-            color: jkl.$color-svaberg;
-        }
-    }
-
-    @include small-device {
-        &__dark-color {
-            margin: 0 -#{jkl.$spacing-m};
-            padding-left: jkl.$spacing-m;
-            padding-right: jkl.$spacing-m;
-
-            &:before {
-                left: jkl.$spacing-m;
-            }
-        }
-    }
-
-    @include from-medium-device {
-        display: grid;
-        grid-template-columns: rem(390px) 1fr 1fr;
-
-        &__info-wrapper {
-            padding-right: jkl.$spacing-xl;
-        }
-
-        &__light-color,
-        &__dark-color {
-            padding: jkl.$spacing-2xl jkl.$spacing-m;
-        }
-
-        &:nth-of-type(1) {
-            margin-top: jkl.$spacing-2xl;
-        }
-    }
-}
