--- portal/src/components/Documentation/Colors/Colors.scss
@@ -1,46 +1,0 @@
-@use "~@fremtind/jkl-core/jkl";
-
-@import "~@fremtind/jkl-core/mixins/_all.scss";
-@import "~@fremtind/jkl-core/_functions.scss";
-
-.jkl-portal-color-info {
-    display: inline-flex;
-
-    &__swatch {
-        width: rem(104px);
-        height: rem(104px);
-        flex-shrink: 0;
-        margin-right: jkl.$spacing-l;
-    }
-
-    &--vertical {
-        flex-direction: column;
-
-        .jkl-portal-color-info__swatch {
-            margin-bottom: jkl.$spacing-xl;
-        }
-    }
-}
-
-.jkl-portal-color-swatch {
-    min-height: rem(20px);
-    min-width: rem(20px);
-    position: relative;
-
-    &__diamond {
-        fill: currentColor;
-        stroke-width: 0.1;
-        stroke: currentColor;
-    }
-
-    &--snohvit {
-        .jkl-portal-color-swatch__diamond {
-            stroke: jkl.$color-granitt;
-        }
-    }
-    &--granitt {
-        .jkl-portal-color-swatch__diamond {
-            stroke: jkl.$color-snohvit;
-        }
-    }
-}
