+++ portal/src/components/colors/ColorInfo.tsx
@@ -1,0 +1,40 @@
+import React, { VFC, useState } from "react";
+import cx from "classnames";
+import { ColorSwatch, rgbToHex } from "./";
+import "./color-info.scss";
+
+export interface Color {
+    colorVariable: string;
+    cmyk?: string;
+    pantone?: string;
+}
+
+interface ColorInfoProps extends Color {
+    vertical?: boolean;
+    className?: string;
+}
+
+export const ColorInfo: VFC<ColorInfoProps> = ({ colorVariable, cmyk, pantone, vertical, className }) => {
+    const [color, setColor] = useState<number[]>();
+    return (
+        <article
+            className={cx("jkl-portal-color-info", className, {
+                "jkl-portal-color-info--vertical": vertical,
+            })}
+        >
+            <ColorSwatch setColor={setColor} className="jkl-portal-color-info__swatch" colorVariable={colorVariable} />
+            <p className="jkl-portal-color-info__values jkl-small">
+                {color && color.length === 3 && (
+                    <>
+                        {`RGB: ${color[0]} ${color[1]} ${color[2]}`} <br />
+                        {`HEX: ${rgbToHex(color)}`}
+                    </>
+                )}
+                {cmyk && <br />}
+                {cmyk && `CMYK: ${cmyk}`}
+                {pantone && <br />}
+                {pantone}
+            </p>
+        </article>
+    );
+};
