+++ portal/src/components/colors/color-info.scss
@@ -1,0 +1,21 @@
+@use "~@fremtind/jkl-core/jkl";
+
+.jkl-portal-color-info {
+    display: inline-flex;
+    align-items: center;
+
+    &__swatch {
+        width: jkl.rem(104px);
+        height: jkl.rem(104px);
+        flex-shrink: 0;
+        margin-right: jkl.$spacing-l;
+    }
+
+    &--vertical {
+        flex-direction: column;
+
+        .jkl-portal-color-info__swatch {
+            margin-bottom: jkl.$spacing-xl;
+        }
+    }
+}
