+++ portal/src/components/colors/color-swatch.scss
@@ -1,0 +1,49 @@
+@use "~@fremtind/jkl-core/jkl";
+@use "~@fremtind/jkl-core/variables/colors";
+
+$farger: (
+    hvit: jkl.$color-hvit,
+    snohvit: jkl.$color-snohvit,
+    sand: jkl.$color-sand,
+    dis: jkl.$color-dis,
+    varde: jkl.$color-varde,
+    svaberg: jkl.$color-svaberg,
+    stein: jkl.$color-stein,
+    fjellgra: jkl.$color-fjellgra,
+    skifer: jkl.$color-skifer,
+    granitt: jkl.$color-granitt,
+    svart: jkl.$color-svart,
+    suksess: colors.varslingsfarge("suksess"),
+    feil: colors.varslingsfarge("feil"),
+    info: colors.varslingsfarge("info"),
+    advarsel: colors.varslingsfarge("advarsel"),
+);
+
+.jkl-portal-color-swatch {
+    min-height: jkl.rem(20px);
+    min-width: jkl.rem(20px);
+    position: relative;
+
+    &__diamond {
+        fill: currentColor;
+        stroke-width: 0.1;
+        stroke: currentColor;
+    }
+
+    &--granitt-stroke {
+        .jkl-portal-color-swatch__diamond {
+            stroke: jkl.$color-granitt;
+        }
+    }
+    &--snohvit-stroke {
+        .jkl-portal-color-swatch__diamond {
+            stroke: jkl.$color-snohvit;
+        }
+    }
+}
+
+@each $property, $value in $farger {
+    .jkl-portal-color-swatch--#{$property} {
+        color: $value;
+    }
+}
