+++ portal/src/components/colors/colors.d.ts
@@ -1,0 +1,1 @@
+export type RgbValue = number[];
