+++ portal/src/components/colors/index.ts
@@ -1,0 +1,7 @@
+export * from "./colors.d";
+export * from "./ColorInfo";
+export * from "./ColorSwatch";
+export * from "./ColorTable";
+export * from "./ColorTableRow";
+export * from "./colorUtils";
+export * from "./PrimaryColor";
