+++ portal/src/components/colors/primary-color.scss
@@ -1,0 +1,29 @@
+@use "~@fremtind/jkl-core/jkl";
+@use "~@fremtind/jkl-core/mixins/_all.scss" as mixins;
+
+.jkl-portal-primary-color {
+    width: 100%;
+    max-width: jkl.rem(768px);
+    margin-top: jkl.$spacing-3xl;
+
+    @include mixins.small-device {
+        &__title,
+        &__swatch {
+            margin-bottom: jkl.$spacing-xl;
+        }
+    }
+
+    @include mixins.from-medium-device {
+        display: grid;
+        grid-template-columns: jkl.rem(340px) 1fr;
+        grid-template-rows: jkl.rem(22px) 1fr;
+
+        &__swatch {
+            grid-row: 1 / 3;
+        }
+
+        &__title {
+            grid-column: 2 / 3;
+        }
+    }
+}
